Update RecyclingInfoScreen and Section components
The changes in this commit update the RecyclingInfoScreen and Section components.

In the RecyclingInfoScreen component, a new import statement is added to import the Ionicons component from the '@expo/vector-icons/Ionicons' package. Two new views are also added to display the date and location information using the Ionicons component for icons.

In the Section component, the MapView component and its related code are removed.

These changes refactor the code to improve the display of recycling information and remove the unnecessary map functionality.