Update RecyclingInfoScreen and Section components
This commit refactors the RecyclingInfoScreen and Section components.

In the RecyclingInfoScreen component, it adds an import statement for the Ionicons component from the '@expo/vector-icons/Ionicons' package. It also introduces two new views to display date and location information using the Ionicons component for icons.

In the Section component, it removes the MapView component and related code.

These changes improve the display of recycling information and remove unnecessary map functionality.